Really fix TestApp_RunAsSubCommandIncorrectUsage
An earlier commit 6564c0f1f9 made an attempt at fixing this test,
but apparently it is not working correctly with Go <= 1.16, because
the panic is expected, and it only happens with Go 1.17+.

Rather than adding an ugly kludge checking the runtime go version,
let's have two versions of the test.

Once Go 1.16 is no longer supported, we can remove the go116_test.go
file.
